requirements.txt
streamlit>=1.28.0
Pillow>=10.0.0
pillow-heif>=0.13.0

